traslado de carpetas de paginas a home, nuevo layout.tsx
