+++ src/app/HOME/layout.tsx
@@ -1,0 +1,14 @@
+import Navbar from "@/components/Navbar/Navbar"
+import Sidebar from "@/components/Sidebar/Sidebar"
+
+export default function HomeLayout({ children }: { children: React.ReactNode }) {
+    return (
+        <div className="flex h-screen bg-gray-100">
+            <Sidebar />
+            <section className="flex-1 p-6 overflow-auto">
+                <Navbar />
+                {children}
+            </section>
+        </div>
+    )
+}
